refactor and unit tests

+++ henripayapi/src/main/java/com/henripay/henripayapi/client/MandateClient.java
@@ -1,6 +1,6 @@
 package com.henripay.henripayapi.client;
 
-import com.henripay.henripayapi.model.MandateDTO;
+import com.henripay.henripayapi.dto.MandateDTO;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.bind.annotation.PathVariable;
 import reactor.core.publisher.Mono;
@@ -8,5 +8,5 @@ import reactor.core.publisher.Mono;
 public interface MandateClient {
 
     @GetExchange("/mandates/{id}")
-    Mono<MandateDTO> getMandateDetails(@PathVariable("id") Integer id);
+    Mono<MandateDTO> getMandateDetails(@PathVariable("id") Long id);
 }
